Backend/Sprint 8/api/requirements.txt
django-cors-headers
django
djangorestframework